Make both job and package optional for each process

+++ lib/nosh.rb
@@ -16,13 +16,17 @@ class Nosh
 
     processes = Dir.glob("#{@component_path}/.nosh/*").map {|path| File.basename(path)}
     processes.each do |process_name|
-      # copy job
-      FileUtils.mkpath("#{@dest_release_path}/jobs/#{@component_name}-#{process_name}")
-      system!("cp -r #{@component_path}/.nosh/#{process_name}/job/ #{@dest_release_path}/jobs/#{@component_name}-#{process_name}/")
+      if Dir.exist?("#{@component_path}/.nosh/#{process_name}/job")
+        # copy job
+        FileUtils.mkpath("#{@dest_release_path}/jobs/#{@component_name}-#{process_name}")
+        system!("cp -r #{@component_path}/.nosh/#{process_name}/job/ #{@dest_release_path}/jobs/#{@component_name}-#{process_name}/")
+      end
 
-      # copy package
-      FileUtils.mkpath("#{@dest_release_path}/packages/#{@component_name}-#{process_name}")
-      system!("cp -r #{@component_path}/.nosh/#{process_name}/package/ #{@dest_release_path}/packages/#{@component_name}-#{process_name}/")
+      if Dir.exist?("#{@component_path}/.nosh/#{process_name}/package")
+        # copy package
+        FileUtils.mkpath("#{@dest_release_path}/packages/#{@component_name}-#{process_name}")
+        system!("cp -r #{@component_path}/.nosh/#{process_name}/package/ #{@dest_release_path}/packages/#{@component_name}-#{process_name}/")
+      end
     end
 
     # copy source
